requirements.txt
streamlit
openai
scikit-learn
numpy
pandas
python-dotenv
fpdf
